Initial Modal move and componentization. Everything's broken right now
