+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/page/workspaces/summary/AclEditor.scala
@@ -1,8 +1,7 @@
 package org.broadinstitute.dsde.firecloud.page.workspaces.summary
 
-import org.broadinstitute.dsde.firecloud.component._
+import org.broadinstitute.dsde.firecloud.component.{OKCancelModal, _}
 import org.broadinstitute.dsde.firecloud.component.Component._
-import org.broadinstitute.dsde.firecloud.page.OKCancelModal
 import org.broadinstitute.dsde.workbench.service.WorkspaceAccessLevel.WorkspaceAccessLevel
 import org.openqa.selenium.WebDriver
