+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/page/workspaces/summary/CloneWorkspaceModal.scala
@@ -1,8 +1,7 @@
 package org.broadinstitute.dsde.firecloud.page.workspaces.summary
 
-import org.broadinstitute.dsde.firecloud.component._
+import org.broadinstitute.dsde.firecloud.component.{OKCancelModal, _}
 import org.broadinstitute.dsde.firecloud.component.Component._
-import org.broadinstitute.dsde.firecloud.page.OKCancelModal
 import org.openqa.selenium.WebDriver
 
 class CloneWorkspaceModal(implicit webDriver: WebDriver) extends OKCancelModal {
